Issue #11446: Updated testHaltOnExceptionOff in CheckerTest
